+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -51,7 +51,7 @@ Potentially Breaking Changes
 
 - Deprecated the :program:`clang-tidy` check :doc:`performance-faster-string-find
   <clang-tidy/checks/performance/faster-string-find>`. It has been renamed to
-  :doc:`performance-faster-string-operation
+  :doc:`performance-prefer-single-char-overloads
   <clang-tidy/checks/performance/faster-string-operation>`.
   The original check will be removed in the 25th release.
 
@@ -197,7 +197,7 @@ New check aliases
 
 - Renamed :doc:`performance-faster-string-find
   <clang-tidy/checks/performance/faster-string-find>` to
-  :doc:`performance-faster-string-operation
+  :doc:`performance-prefer-single-char-overloads
   <clang-tidy/checks/performance/faster-string-operation>`.
   The `performance-faster-string-find` name is kept as an alias.
 
@@ -341,7 +341,7 @@ Changes in existing checks
 
   - Improved the ignore list to correctly handle ``typedef`` and  ``enum``.
 
-- Improved :doc:`performance-faster-string-operation
+- Improved :doc:`performance-prefer-single-char-overloads
   <clang-tidy/checks/performance/faster-string-operation>` check:
 
   - Now analyzes calls to the ``starts_with``, ``ends_with``, ``contains``,